+++ packages/@dicebear/shapes/src/components/index.ts
@@ -1,0 +1,9 @@
+export { centerModifier } from './centerModifier';
+export { cornersModifier } from './cornersModifier';
+export { sidesModifier } from './sidesModifier';
+export { sidesWrapper } from './sidesWrapper';
+export { sides } from './sides';
+export { cornersWrapper } from './cornersWrapper';
+export { corners } from './corners';
+export { centerWrapper } from './centerWrapper';
+export { center } from './center';