+++ packages/@dicebear/shapes/src/index.ts
@@ -1,0 +1,18 @@
+/*!
+ * Shapes (@dicebear/shapes)
+ *
+ * Code licensed under MIT License.
+ * Copyright (c) 2014-2021 Daniel Mester Pirttijärvi
+ * Copyright (c) 2021 Florian Körner
+ *
+ * Design "Jdenticon" by Daniel Mester Pirttijärvi licensed under MIT.
+ * Source: https://github.com/dmester/jdenticon
+ * License: https://github.com/dicebear/dicebear/blob/main/packages/%40dicebear/shapes/LICENSE
+ */
+
+import { style } from './core';
+
+let { create, meta, schema } = style;
+
+export { create, meta, schema };
+export { Options } from './options';